We now have working level switching and AMenu/AGeometry are their own QIrrWidgets.
Level Switching:
Call ABase::changeToLevel and pass as argument the level you want to switch to. For AMenu it is "menu" and for AGeometry UI it is "geometry.ui". The .ui means that we do not need to setup any special signals, so only automatically load the .ui file.

AMenu:
This was quite simple to severe from ABase, because it only used ABase's drawing functions. Those now belong to QIrrWidget. However the menu does not work fully, instead it ends at a wrong position (unless you skip the animation with SPACE). I am guessing it has to do with the timer being set long before the execute functions are called, I will invistage this next.

AGeometry:
A bit more complicated, because ABase also had some tour stuff. I copied those into the AGeometry class, so any tour/event classes think it is still ABase.