I forgot to commit a few files last time: 	- New button based menu with connection to AMenu 	- The geometry UI uses QIrred AGeometry 	- Old event receiver removed from Makefile, because events are now handled by the QIrrWidgets
Also I tried to find the problem with AMenu and while looking around I noticed the 
tripod menu did not end up in the location it originally started in, like rest of 
the animated letters. Since this was a result of +/- being switched and inconsistent 
with rest, so I figured it was a typo. After that I could not reproduce the loading 
problem I had with AMenu, so I decided to commit it too.