For some reason the AGeometry FPS camera could not receive any input starting today. Reverting to older AMELIA/IrrLicht did not help, nor did doing a clean check out of both. What makes it weirder is that Jacob did not experience any problems with the same revisions of both.
However making sure that AGeometry::OnEvent returned something seemed to fix this. Since the declaration of OnEvent specifies that a boolean should be returned (but none was in the implementation), my fix seems like a good idea to commit in order to ensure that the return value is always what we want it to be.

Note: IrrLicht documentation says that returning false will keep the event propagating down the chain of event receivers and returning true will stop the event from propagating.