Scaling of the viewport images implemented.  Now createImageWithOverlay takes two more arguments which specify the rectangle into which each of the two images should be scaled into. They have a default value of null rectagle, in which case the rectangle is found with the QImage::rect() function.
The last problem left is that when you click the viewport, the Qt button background is still being drawn which you can see through the transparent edges. While it is not being clicked, the background is not being drawn because I set the QPushButton::flat property to true. I am playing around with a fix now, but if it does not work we can always set the background image of the viewports to a transparent PNG in the stylesheet.