Added a new track type, ARTrack, describing reconstructed tracks. Added a new node type ARTrack3DNode for the graphical representation of ARTracks
The loading of tracks is working fine, but the scale factor and covariant matrix should be adjusted. No selection or combination of tracks yet.
This is a HUGE can of worms. Since there's no clear information about what's on each track, a lot of the functionalities of AMELIA will need to be put in perspective again.